Adds collection FILES_TMP_CONTENT
The collection is filled with a key-value set where value is the content of
the file which was uploaded. This collection can be used with all supported
operators, however, SecTmpSaveUploadedFiles should be set to 'On' in order
to have this collection filled. Note that @inspectFile is now depending on
SecTmpSaveUploadedFiles. This is necessary to keep performance while such
functionalities where not used.