Increases the loglevel while failed to delete a collection
One of the reasons that sdbm failed to delete a key is the fact that it may
already been excluded). The error logs are not clear regarding that. Check
if the key exists prior to delete will generate unecessary computation. This
commit increases the loglevel of the error message and makes it acts like the
operation was successfully accomplished.