Fix cases when shared memory allocated by first instance of application being overwritten by another instance attaching same shared memory segments.  It happens in case when 'nginx' server is running, but configuration is updated and 'nginx -t' is called in order to update configuration.  Shared memory segment gets overriten by second instance of nginx, and any call to LogFile/SharedFile::find_handler() will cause crash.
With this fix, every initialized block is being marked with feature specific signature
assuming that block in use already, and can't be re-used yet.  For second and later
application instances yet another ftok() code is adviced to use (via recursive call to
add_new_handler()).

There's a side effect possible when nginx (or other user of libmodsecurity) gets
crashed, shared memory block can be left in place with valid signature stored in there.
For this case it's advised to use ipcs(1)/ipcrm(1) to cleanup stale blocks.