Add neutralize safe encoding transformation
Many times we have false positive due to random pattern in string encoded
in base64 or hexadecimal. This patch propose a transformation which escape
known and valid encoding. If the string to be transformed represent from
the first character to the last a valid encoding, all the characters are
replaced by a tag describing the escaped encoding, and fill other
characters by "a".

the recongnized encoding are:

 - base64 (only [a-zA-Z0-9+/] and eventually "=" at the end, and a
           multiple of 4 characters)

 - base64i (only [a-zA-Z0-9-_])

 - hex32 (only [0-9a-f] and multiple of 2 characters)

 - HEX32 (only [0-9A-F] and multiple of 2 characters)

If none of this known encoding is matched, the transformation does nothing.
The inut string is given as-is at the output.

Example:

   input string "Fk2SELECTIUaEGsgww==" containing random word "SELECT"

Becomes

   "b64aaaaaaaaaaaaaaaaa"

And no false positive are triggerred.