Add configure option to parameterize Modsecurity/libcurl timeout for retrieving rules.
- There are scenarios where the internet connection is not available for the httpd,
for example during the test of Disaster Recovery or other failures. In these cases,
it would be great to have a possibility to start the httpd in the same (more or less)
timing of when there is an internet connection. Moreover, being the currently timeout
a default, it is not possible to control this value.