Test for proving Loading rule is not thread safe
When multiple threads are trying to load rule at the same
time even when they refer to different rule set a string
representing rule is malformed. It's believed that
this is caused by seclang-parser which contains plenty
of global variables.

To run the test simply enter 'test' directory and run
"./unit_tests"