Add editorconfig to help OSS contributors
When switching between many different open source code bases, it can be
cumbersome to constantly change editor settings in order to respect a
given project's basic style preferences. EditorConfig is a popular
format for representing those preferences so editors can automatically
respect them.

Many editors support EditorConfig out of the box, and many others have
EditorConfig plugins:
https://editorconfig.org/#pre-installed

This commit adds an EditorConfig to the ModSecurity project. It is most
likely not perfect because the chosen indentation varies a bit
throughout the project, but hopefully it provides a good basis for most
ModSecurity editing.