test: don't directly use the .libs directory
The build system should not directly use the .libs directory which is
intended for internal use by the libtool implementation. Additionally
when linking with internal a library the .la file should be used as this
provides all the information required by the libtool implementation.

When using Slibtool instead of GNU Libtool the build fails because
Slibtool is more strict about this and fails to find the required files
to link with.